homework refactor

+++ src/HomeWork/L3/T1_Calculator.java
@@ -8,13 +8,13 @@ public class T1_Calculator {
         double result = 0;
 
         System.out.print("Input first number: ");
-        double num1 = scan.nextInt();
+        double num1 = scan.nextDouble();
 
         System.out.print("Choose the action: ");
         String action = scan.next();
 
         System.out.print("Input second number: ");
-        double num2 = scan.nextInt();
+        double num2 = scan.nextDouble();
 
         switch (action) {
             case "+" -> result = num1 + num2;
